Examples update (#99, #62):
- Example 1 is now the zoom one (was: number 2), and I beefed it up
to have the needex HTML markup as well and have a takePhoto().
- Example 2 is now the repeated-grabbing-of-a-frame, no big changes,
just homogeneisation of variable names.
- Example 3 is now the grab-frame-and-postprocessing, which I made
work by bouncing the grabbed `imageBitmap` on a `canvas`. For the
purpose of the example, the colours are inverted.
- Also I homogeneised catch/then clause indenting and error reporting,
and inlined the superfluous `failedToGetMedia()`.

Finally I also added 3 codepen.io links (and a collection) with each
of the examples slightly modified, and added informative notes with
the links to the spec. Slightly modified means I added buttons to e.g.
start/stop capture and sometimes I casted the video feed to a video tag.