Add tag-security-privacy.md
Add the tag-security-privacy document to assist in the TAG review.
This was requested in the [Intent to Ship thread](https://groups.google.com/a/chromium.org/d/msg/blink-dev/j-Q08QgBipM/RgGOQxG-BwAJ)

This document right now covers Pan/Tilt but I guess is applicable to
most of the other camera properties on this specification.